[RUN3, PWGGA] Fix circle dependency
The `AliAnalysisTaskAO2DConverter` needed to know of `AliV0ReaderV1` and vice versa.
This commit adds the class `AliAODRelabelInterface` which is a new base class that inherits itself from the AliAnalysisTaskSE class and that AliV0ReaderV1 and AliAnalysisTaskAO2DConverter inherit from. This way both classes only need the one new class and do not need to know from each other.
Inside PWGGA/GammaConvBase/AliConversionPhotonCuts both classes were also needed because there in the function `GetTrack` one needs to check if the V0s are already relabeled.